requirements.txt
beautifulsoup4==4.9.3
bs4==0.0.1
python-dotenv==0.15.0
requests==2.25.1
urllib3==1.26.3
